requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
Django==3.0.8
django-allauth==0.42.0
django-braces==1.14.0
django-cors-middleware==1.5.0
django-extensions==3.0.2
django-filter==2.3.0
django-guardian==2.3.0
django-model-utils==4.0.0
django-oauth-toolkit==1.3.2
django-rest-auth==0.9.5
djangorestframework==3.11.0
idna==2.10
oauthlib==3.1.0
psycopg2-binary==2.8.5
pycparser==2.20
PyJWT==1.7.1
python-http-client==3.2.7
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
sendgrid==6.4.1
six==1.15.0
sqlparse==0.3.1
starkbank-ecdsa==1.0.0
uritemplate==3.0.1
urllib3==1.25.9
uuid==1.30
whitenoise==5.1.0
